v2c-python-flask/pyproject.toml
[tool.poetry]
name = "v2c-python-flask"
version = "0.1.0"
description = "Reference implementation of v2c"
authors = ["inqb"]
license = "APACHE-2.0"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
Flask = "^1.1.2"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.5"
black = "^20.8b1"
reorder-python-imports = "^2.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
